requirements.txt
pandas>=2.0.0
numpy>=1.24.0
torch>=2.0.0
scikit-learn>=1.2.0
plotly>=5.14.0
pyyaml>=6.0.0
fpdf>=1.7.2
openpyxl>=3.1.0
streamlit>=1.20.0
prophet>=1.1.0
kafka-python>=2.0.0
great-expectations>=0.15.0
transformers>=4.30.0
translate>=3.6.0
schedule>=1.2.0
slack_sdk>=3.20.0
tenacity>=8.2.0
watchdog>=3.0.0
prefect>=2.10.0
psutil>=5.9.0
prometheus_client>=0.17.0
structlog>=23.1.0
fastapi>=0.100.0
uvicorn>=0.23.0